requirements.txt
streamlit>=1.28.0
flask==3.0.2
flask-cors==4.0.0
google-cloud-storage==2.14.0
python-dotenv==1.0.1
pandas>=2.0.0
plotly>=5.15.0
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
requests==2.31.0
pillow>=9.5.0
